pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-trend-predictor"
version = "0.1.0"
description = "Minimalistic ML pipeline for predicting whether a news item will trend."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4,<4.0.0",
  "gspread>=6.0.2,<7.0.0",
  "joblib>=1.3.2,<2.0.0",
  "numpy>=1.26.4,<3.0.0",
  "pandas>=2.2.2,<3.0.0",
  "pydantic>=2.7.1,<3.0.0",
  "pydantic-settings>=2.2.1,<3.0.0",
  "python-dateutil>=2.9.0.post0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "requests>=2.32.3,<3.0.0",
  "scikit-learn>=1.5.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
